renamed and moved helper files for clarity
