+++ internal/framework/mount/mount_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package sys
+package mount
 
 import (
 	"context"
